requirements.txt
requests
beautifulsoup4
lxml
nltk
spacy
sentence-transformers==2.2.2
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain-google-genai
langchain_experimental
langgraph
SpeechRecognition
pyttsx3
pyaudio